Fix require('mock-modules').dumpCache() when a global var isn't an object
This is a dumb no-brainer -- currently dumpCache() iterates through all the
global vars and calls .mockClear on them if they have a mockClear method.

However, things that aren't objects don't have methods...so we need to be
sensitive to that when checking to see if a global object is a mock that needs
to be mockClear()ed.