Abstract most of testRunner.js
Pulls out most of the guts of testRunner.js and puts it into a reusable library.
Every project should have its own runner CLI, but every project shouldn't have
to duplicate all that logic.

This diff moves almost all of it into a TestRunner class that a CLI can just new
up and call .run() on