Fix test-execution errors + make TestRunner print relative test path
This fixes a bug where, if an exeception was thrown at the top-level of a test
file, the error would propogate up through the worker pool without being caught
and stop the TestRunner from continuing on with other tests. Obviously we should
just mark the test as failed, print the error message, and move along to any
remaining tests.

Additionally this makes TestRunner print out paths relative to the rootDir
specified in the config (if present), rather than the full absolute path to the
test. This just makes the console output a little less redundant.