Add support for running tests in-band (i.e. not via a worker pool)
This refactors TestRunner.js and TestWorker.js a bit to make it possible to run
tests in-band in the current process. This is really useful when debugging test
failures because, when an error occurs somewhere in the internals of jest, it
becomes easier to extract traces find errors in the current process rather than
hoping that the error information is proxied through to the parent process.

This should also make it simpler to integrate node-inspector with jest.