Ensure generateMock() can't cause side-effects on the real module environment
generateMock() executes the module for which a mock is being generated (in order
to examine its exports and build a corresponding mock). However, if the factory
for the mock under examination (or the factory for any of its unmocked
dependencies) does something that causes a side-effect in the module system,
this side-effect would be observable. This is really bad!

So to ensure this can't happen, this diff temporarily shims the module and mock
registries while executing the module so that all side-effects that may occur
can be thrown away.