Fix issue with require('mock-modules').setMock() when manual mock exists
Previously if you called setMock('Foo', FooExports), FooExports would be placed
into the mock registry by resolving the real modulePath for 'Foo'.

However, if someone were to later requireMock('Foo') (and a manual mock for Foo
existed), requireMock would resolve the path to the manual mock first and thus
not see the explicitly set mock in the mock registry.

This fixes the issue by keeping a separate explicitly-set mock registry that is
checked in requireMock before anything else.