Normalize module names across relative paths, @providesModule names, and manual mocks for all mock-modules util functions
Previously if you were to change the mocking status for a specific module in one
of the mock-modules APIs (like dontMock() or setMock()), that change would only
be recorded for the form of the module name you gave.

For example:
require('mock-modules').dontMock('../myModule');
require('../myModule'); // Would not be mocked
require('/absolute/path/to/myModule'); // Would still be mocked!

This adds a new method that will normalize any (currPath, moduleName) tuple into
a string that is the same for all of: The moduleName, any path to the module,
any path to the module's manual mock (if a manual mock exists).