Add a hack for node-haste where package.json package sub-module ids ind with '.js'
This is a hack for an issue that should really just be fixed in node-haste.
The only reason I'm adding the hack here for now is because updating node-haste
for this is a bit complex at the moment given that www seems to have diverged
from open source slightly.

Will come back to this and fix it properly in node-haste later on