Refactor TestRunner.runAllMatchingParallel/InBand
This replaces these two APIs with runTestsParallel and runTestsInBand that
accept an array of exact paths (rather than a RegExp pattern).

Since TestRunner.findTestPathsMatching() is a public API, the idea is that
external callsites should use that directly to reduce a pattern to a list of
path strings. Then, from there, pass that list of path strings in to the
concrete path APIs.