Fix node 0.8.8 segfault issue
It appears to be the case that, in node 0.8.8 (haven't tested other versions
yet), a worker that has executed a large number of tests can ultimately wind up
segfaulting when we .close() the jsdom contextify instance.

I have no idea what's happening -- but the repro was consistent, and stops as
soon as we stop closing the contexts.

Will need to come back to this and understand it better...but for now we need to
stop the bleeding.