Make bin/jest.js point directly at CWD-local "jest" bin (rather than node_modules/.bin/jest)
npm on windows doesn't use some form of symlinks like on *nix, but instead uses
wrapper batch files of some kind. As such, it's unreliable to assume
that node_modules/.bin/jest is a JS module.

This changes bin/jest.js to, instead, try to directly require the binary
path.