Print pending test info when no test result activity has happened in a while
Currently, if you're running a large number of tests and one of those
tests is very slow, sometimes it's possible that Jest looks like it has
hung (when really it's just waiting for that slow test to finish).

This adds a feature to the test runner that prints/logs out pending test
info if no test results have come back for a while (5 seconds, for now).
Hopefully this should help give better feedback to users as to what's
going on when this happens.