Eliminate dependency on WorkerPool.sendMessageToAllWorkers()
That API has been blocking node-worker-pool from being able to recover
more gracefully when a child process goes awry (seg faults, OOMs, etc).

As long as that API exists, node-worker-pool must guarantee that all
workers have received some message. If that is the case, then
node-worker-pool can't restart a worker (with the assumption that it is
stateless) because the newly booted replacement worker process will not
have received the message all other workers did some time in the past.

It turns out we don't really need to send a signal to all workers when
the resource map has been written because we don't even start sending
messages to workers until it's been written. So this just makes the
worker aware of the fact that it needs to read the resource map on the
first message it receives before continuing on to build a response.