Speed up output for failing tests
Summary:

Jest may appear to hang (for dozens of seconds, up to minutes) when
printing failed test results.

An example of the kind of failure output that can trigger this condition
is a dump of a very large JS object and a short string. All of the time
is spent in the diffing algorithm that we use to color-highlight the
differences between the expected and actual value. This scenario is one
of the worst cases for the underlying Myers LCS diff algorithm, which
runs in O(n+d^2) time (where "d" is the length of the edit script
between the two strings).

This commit ameliorates the worst case scenario by switching to a
line-level diff whenever newlines are present in either string. There
will still be pathological cases that can be very slow (really huge
multi-line diffs, or very long single-line diffs) but this change is
relatively cheap and delivers a nice benefit for the most common cases.

Reference: D1896663

Test Plan:

Add automated tests (unfortunately, fairly tied to Jasmine specifics)
and run using `npm test`.

In a separate project, add some test failures, see fast line-level diffs
for examples like the one mentioned above, and character-level diffs in
other cases.