Fixing code coverage in Windows
Transform module paths/RegExps so the path matching logic works on Windows.
Use path.delimiter instead of ":" as a separator in module IDs, since ":" is a valid path character on Windows... All module paths in Windows were getting mapped as {drive letter} which wasn't super-useful.
Changed a check for absolute paths (check that first character is "/") to a more reliable cross-platform check for absolute paths (use path.isAbsolute).
Updated tests to expect different paths on different platforms.
Updated a unit test in HasteModuleLoader-requireMock-test.js to look for a Windows error string as well as the Unix string