[jest] Elevate jasmine-pit into jest
Summary: I merged jasmine-pit into jest because there is really no point in having it in a separate repo. If we decide to make jest more modular we should make `jest-jasmine` its own package within jest, which will allow us to move faster. This is the same that we did for jasmine-only.
I lifted the implementation from gagoman's pull request ( https://github.com/jeffmo/jasmine-pit/pull/12 ) into the repo and modernized it for ES2015. (Thanks gagoman for working on this! All the credit goes to you!). This is a precursor to adding jasmine 2 support.

@public

Reviewed By: yungsters

Differential Revision: D2662889

fb-gh-sync-id: 220eb9f6d9699ac7924bf7c6d8b055d7273cc104