Add a friendly "no tests found" message
Summary:
This adds a message when no tests are found because the pattern specified after the Jest command did not match any filenames.

The message displays the string entered by user. However if that string is different from the regex pattern (for example, "foo bar" becomes /foo|bar/), we also display the regex to make it clear why the pattern didn't match.

We take care to use quotes when displaying user input as a string, and slashes when displaying a regex, so error message stays precise and accurate.

public

Reviewed By: cpojer

Differential Revision: D2823139

fb-gh-sync-id: 8d1b5a0682ee950cc129976f09b824c0c82495f2