Fix VerboseLogger throwing when there is a syntax error
Summary:
The `VerboseLogger` currently throws when there is a syntax error in the test file. This happens because it expects `testResults` to be an array. This is normally the case whether the tests succeed or fail:

{F59917214}

However if there is a syntax error in the test file, a different code path executes in `TestRunner`. This code path provides a different `testResult` object that doesn't look like anything you'd normally receive (all fields are different). Some of these fields make sense in the error context (such as `testExecError`) but some are empty and unused (`tests` and `suites` aren't referenced anywhere else in the project) and some are breaking things (`testResults` is an empty object rather than an array).

{F59917223}

Removing the extra empty fields (`suites` and `tests`) and changing `testResults` to be an array fixes the crash in `VerboseLogger` and doesn't appear to affect anything else. After this fix, the corrects stack trace is printed at the end of the verbose test run.

Later we may want to enhance the in-progress view of the verbose test run, as currently we only provide details about the syntax error at the very bottom. This can be done separately.

public

Reviewed By: cpojer

Differential Revision: D2840519

fb-gh-sync-id: 2133d41a3ad1db72c454e3fe698297eb7a561e60