Add babel-plugin-jest-unmock and babel-preset-jest.
Summary:This adds two new packages as part of Jest's modularization effort. One of them is a babel plugin to hoist calls to the new `jest.unmock` API in order to make `import` statements work properly with unmocking behavior. The old API `dontMock` will not be hoisted and serve as an escape hatch. Note that this transform hoists every `jest.unmock` call to the top of their scope. I will also update the documentation (separately) to highlight this new feature for the new 0.9.0 release. I will pull in `babel-jest` into the jest repo and make this transform a default in a follow-up diff as well. If you'd like to try out this feature immediately, install `jest-cli@next` and add `babel-preset-jest` to your `.babelrc` file.

See http://astexplorer.net/#/SSAVF697Ro/4

cc davidbarton sameoldmadness jeffcarp
cc kittens
Closes https://github.com/facebook/jest/pull/741

Reviewed By: kentaromiura

Differential Revision: D2976284

fb-gh-sync-id: 76d0745c230916f71da3b393d3d3698876f77134
shipit-source-id: 76d0745c230916f71da3b393d3d3698876f77134