Initial implementation of a blog support
This adds blog post rendering, left nav with the list of blog posts, highlighting the currently selected blog post and highlighting the header links.

I cp all the React blog posts in order to test it out, working great.

Missing pieces:
 - I haven't implemented a listing of all blog posts, the header link opens the most recent one.
 - I haven't implemented truncation of the links on the left and the all page

Those are needed when we'll have > 1 blog post, which is not now :)

@bypass-lint
(for the gitignore)