Update babel-jest for better code coverage
Summary:The current configuration of babel-jest does not allow for code coverage to get close to 100 because Babel writes functions into the compiled source code. By utilizing the babel-plugin-transform-runtime we are able to tell Babel to require these functions from babel-runtime instead. We are also adding a comment that tells babel to ignore Babel generated lines. In most cases this will allow us to get to 100 coverage.
Closes https://github.com/facebook/jest/pull/785

Differential Revision: D3037359

fb-gh-sync-id: 2cf83875dbdfd49fe46ad4094710c1a014d5c101
shipit-source-id: 2cf83875dbdfd49fe46ad4094710c1a014d5c101