Improve `--watch` performance and fix `jest -o`.
Summary:* We now limit the default number of CPUs to `cpus - 1`. This doesn't have a very noticeable impact on performance because of all the caching we are now doing in workers
* For `--watch`, we are now using `floor(cpus / 2)` so that people can comfortably keep using their editors.
* When `jest -o` was invoked with no changed files, it would hang. Fixed that and added a descriptive message.
* Removed some outdated and unused config options for TestRunner.
* Bumped the timeout for the jsdom test because travis sometimes takes too long.

The long term plan is to split the TestRunner code into a simple TestRunner and a class that finds matching tests. That should clean everything up quite a bit :)
Closes https://github.com/facebook/jest/pull/786

Reviewed By: yungsters

Differential Revision: D3037390

fb-gh-sync-id: a29882828d057ff7a1850255072e367ed0d28e16
shipit-source-id: a29882828d057ff7a1850255072e367ed0d28e16