Rename pending to skipped, fix “Running XX tests” message.
Summary:`pending` is confusing to people at FB and I agree. Skipped makes more sense. This only updates the user output and continues to use `pending` in code – to align it with what Jasmine uses internally.

I also removed a reference to a variable that doesn't exist which broke the `Running XX tests...` feature.
Closes https://github.com/facebook/jest/pull/820

Reviewed By: vjeux

Differential Revision: D3079730

fb-gh-sync-id: f19e30f6f24095ff45563c5f0dfc957f8a08a2dc
shipit-source-id: f19e30f6f24095ff45563c5f0dfc957f8a08a2dc