Add developer environment setup script and fix symlink issues.
Summary:This diff adds a script after `postinstall` that sets up the dev environment using `lerna`. Once we move jest's source itself into a package, we can drop the custom calls to `npm link`.

Linking `jest-util` pointed out a symlink unmocking bug in Jest I've been meaning to fix for a long time. After running the setup script, Jest would not be able to mock `chalk` properly if `jest-util` was linked into Jest. It turns out we are calling `fs.realpathSync`. I think this was necessary with for `node-haste1` support and is a leftover, however symlink unmocking was never properly supported then so I'm questioning why it was there in the first place.
Closes https://github.com/facebook/jest/pull/821

Reviewed By: vjeux

Differential Revision: D3080727

fb-gh-sync-id: fd3780f9c30031eb2505c26fb5c944de3f7c948f
shipit-source-id: fd3780f9c30031eb2505c26fb5c944de3f7c948f