Add title and description to documentation pages
All documentation pages currently use the default title which is "Jest |
Painless JavaScript Unit Testing". This isn't very helpful for anyone,
so I am supplying the content's title to the Site so we have something
more useful. I also noticed that blog posts grabbed the first line of
the content for the description, so I included that for free here as
well.