Put content title before site title in website
If you have a lot of tabs open, the titles get truncated. The favicon
should be enough to signify what site you are on, so it is best to put
the most useful information at the beginning of the title. In this case,
it is the title of the content you are viewing and not the site. This
changes tabs from all reading like "Jest | ..." to reading more like
"Tutorial...", "Module Te...", and "Automatic...".

We probably don't even need the site name in the title for these pages
at all, but I decided to keep this change more minimal.