add support of mockImplementationOnce to mock functions
Summary:It would be great to have a way to define multiple mock implementations of a function similar to the way it's done for the `mockReturnValueOnce` method.

It's useful when one needs to simulate a certain behaviour of a function that is called multimple times and has a callback as one if its arguments.

Example:

```javascript
const mockFn = moduleMocker.getMockFunction();

mockFn.mockImplementationOnce((cb) => {
  cb(null, 'Foo');
}).mockImplementationOnce((cb) => {
  cb(null, 'Bar');
}).mockImplementation((cb) => {
  cb(new Error('FooBar!'));
});
```
Closes https://github.com/facebook/jest/pull/866

Differential Revision: D3167703

fb-gh-sync-id: caa282ae96fb214be4f852fa998520bfd09aa8a9
fbshipit-source-id: caa282ae96fb214be4f852fa998520bfd09aa8a9