Update eslint rules and clean up code.
Summary:We spend too much time reviewing minor nits and it increases turnaround time. This updates the eslint config to error on a few stylistic things and it is now aligned with the Jest code style.

This was also a good change (or well, I was forced :P ) to clean up the last of the cruft of the codebase :)
Closes https://github.com/facebook/jest/pull/892

Reviewed By: vjeux

Differential Revision: D3178752

fb-gh-sync-id: 7d1b759dffeca9374ecab39f7f031cf0da6832e3
fbshipit-source-id: 7d1b759dffeca9374ecab39f7f031cf0da6832e3