Add proper support for module.paths.
Summary:Fixes #898. Eslint is using private module APIs, so let's add them to Jest.

We are using a private API from the `module` module (heh!) so we'll guard it with a try-catch in case it ever breaks. We could also lift the API into Jest itself but I don't see a point in recreating it. I added a test that will undoubtedly fail if the API starts breaking in node and when that happens, we can fix it accordingly :)
Closes https://github.com/facebook/jest/pull/910

Reviewed By: kentaromiura

Differential Revision: D3190210

fb-gh-sync-id: 928e4e9adf9aeaabe3a8de02e817118597474aab
fbshipit-source-id: 928e4e9adf9aeaabe3a8de02e817118597474aab