Implement `jest-haste-map` instead of `node-haste`
Summary:This is a new haste map implementation for Jest which is much more scalable than node-haste.

node-haste2 isn't well designed and not scalable for short-lived services like Jest. The startup time of node-haste1 vs. node-haste2 on www is almost the same, both between 6 and 8 seconds which is not acceptable for our engineers. This implementation is attempting to accomplish a much reduced and more scalable startup time. It also has reduced scope ? the goal of this is to only build a haste map and provide a way to resolve a one-level deep dependency tree, which is all that Jest really needs from node-haste. `jest-haste-map` can serve as an ideal basis for rewriting node-haste2 (into node-haste3!).

With this, the cold start time (building the entire haste map) is now about 14 seconds on www (that's acceptable) but the incremental invocation is only 2 seconds (kyldvs will love me). I haven't heavily micro-optimized the JavaScript in `packages/jest-haste-map/src/index.js` and there is a bunch of data copying
Closes https://github.com/facebook/jest/pull/896

Reviewed By: kentaromiura

Differential Revision: D3206943

fb-gh-sync-id: 247c4e15ef6dc7e0b1151b7e2539d053389bd249
fbshipit-source-id: 247c4e15ef6dc7e0b1151b7e2539d053389bd249