Install instructions front and center
I think for lots of folks installing the package is the first thing they think of when trying something out. Personally I kept missing the actual npm i line when scanning the docs. It's frustrating feeling loke Jest is telling me how to use it but being coy about how to get it :P