Install instructions front and center
Summary:
I think for lots of folks installing the package is the first thing they think of when trying something out. Personally I kept missing the actual npm i line when scanning the docs. It's frustrating feeling loke Jest is telling me how to use it but being coy about how to get it :P
Closes https://github.com/facebook/jest/pull/957

Differential Revision: D3240711

fb-gh-sync-id: 571fedfa5c42104f672dea7693cc236756b55118
fbshipit-source-id: 571fedfa5c42104f672dea7693cc236756b55118