Fix jest-haste-map on vagrant.
Summary:
This is a similar fix to what spicyj did in https://github.com/spicyj/jest/commit/cff14103b247fcf539b5d5434b07fb2a55819d6d a long time ago. I wasn't aware of why we had this but it seems obvious to me that we shouldn't create processes if there are less than one worker available. Not really sure how that could happen but it seems that on vagrant/docker installs the number of cores is 0.

See #980.
Closes https://github.com/facebook/jest/pull/984

Reviewed By: dmitriiabramov

Differential Revision: D3279511

fbshipit-source-id: 2e1c939aa09734fe6b00f8169986ca41da2c50cf