Prevent workerCount == 0 on single CPU machines
When the workerCount is not explicitly set by the user, it defaults to the
number of CPUs on the machine, less 1. The intention of
packages/jest-cli/src/TestRunner.js#L46 is to prevent the worker count from
being 0 if it is running on a single-core CPU. However,
packages/jest-cli/src/TestRunner.js#L61 was referencing the passed-in worker
count rather than the sanitized value. This causes single-core machines to be
completely unable to run tests unless they specify the worker count manually.