toMatchSnapshot implementation
getMatchers returns the toMatchSnapshot implementation,
the logic is the following:

if no file exists, it creates the snapshot file and pass the tests;
if a file exists then checks if a specific test exists in the snapshot,
if not it adds it and pass the test.
if both the file exists and already contains a saved snapshot then
checks if it matches the current serialization and return the
appropriate state of the test, if jest is called with
--overwriteSnapshot this last check always returns true and the
snapshot gets updated instead.

In case of differences between the stored snapshot and the one
currently testing it use our formatterMatchFailure to show the
highlighted diff