Move integration tests / update runTests.
Summary:
Move back to `__tests__` files and ignore them in the top level of the integration tests folder. This is so much better!

Also, we are now running the package tests first – because jest-cli is a package, we can just do this and it will run as part of the packages. This should increase the signal while developing.

cc dmitriiabramov
Closes https://github.com/facebook/jest/pull/1002

Reviewed By: dmitriiabramov

Differential Revision: D3295955

fbshipit-source-id: 5e677ee909c2f99820d723f44b7f7ee0a5db45f0