Rewrite of TestSnapshot
Moved almost all file/path access to it (todo moving the last in index),
Create a snapshotState on jest-jasmine2 that gets passed to the
`getMatchers` function, so that common can be deleted.
`getMatchers` uses the snapshotState to gather the Spec fullname and to
increment a counter (I may use an autoincrement getter here in the
future, so that even this state is not handled here).
`initialize` is now `getSnapshotState` which returns the state
management relative to a single file,
TestSnapshot is fully rewritten to load and save a file only once.