Complain loudly about unrecognized args
Jest currently accepts any args at all. Given:

```
jest --silent --invalid foo
```

It will recognize `silent` as a boolean, take `invalid` as an option with value
`foo`, and then run the entire test suite. I noticed this when trying to use
`--silent` on an older version of Jest that doesn't support it:

```
jest --silent filter
```

I had expected `filter` to reduce the number of tests run, but it ran them
all, and on further inspection I discovered that optimist was jamming all the
unrecognized options inside `argv` without complaint.

So, this commit makes it search `argv` for unrecognized keys, and if it finds
any, it bails. Allowed keys are:

- Explicitly defined options.
- Any `alias` set on an explicitly defined option.
- The special `$0` option which holds the original command.
- The special `_` value which holds all non-option arguments.