RFC: switch from optimist to yargs
optimist is old, yargs is new.

optimist truncates the help output in the event that `check()` throws an error,
meaning the user probably doesn't have any idea what went wrong or why.

yargs shows the untruncated help output.

This is an RFC because we run Jest in a bunch of places and not all will have
access to yargs.