Make `Resolver.resolveModule` error node-compliant
Summary:
When you `require()` a non-existing path/module, Node.js will set the
`code` property of the thrown error to `'MODULE_NOT_FOUND'`.

This change ensures that jest behaves the same way so that other code
relying on this specific value of the `error.code` continues to work.

For the Node.js module/error specification, please see:
https://github.com/nodejs/node/blob/v6.2.0/doc/api/modules.md#file-modules
Closes https://github.com/facebook/jest/pull/1061

Differential Revision: D3341023

fbshipit-source-id: 4751f49c70bee1220e7912037527ccbdd440dec9