Snapshot functionality
Summary:
This commit introduce a new matcher to `jest`:
`toMatchSnapshot`.
this new matcher can be called as usual:
`expect(something).toMatchSnapshot()`
the main idea is that the first time the test run it will store whatever is inside expect in a snapshot file,
the next time `jest` runs it will check against it to see if something has changed.

In order to store something on file we need to serialize it, so I currently implemented two different serialization strategies, if we are _expecting_ a `React` element to match a snapshot representation then it will be _currently_ be serialized using the `React.renderToString` method, otherwise it fallbacks to `jest`'s pretty printer.
I choose it because it provides out of the box support for `Maps` and `Sets` other than `Object`s and `Array`s.
As the title mention this is a work in progress as I plan to iterate on it to improve the serialization strategies and possible the output when it fails the expectation.

The matcher uses jasmine `Spec` `getFullName` as the
Closes https://github.com/facebook/jest/pull/1000

Differential Revision: D3341276

fbshipit-source-id: ffff2326c104f26b612fbc4bc6fd2b7dd22f2a31