Add documentation about full paths for *IgnorePatterns options
I was recently bit by a problem where my tests were running perfectly
locally, but were passing in Travis CI without running any tests. I
traced this down to the following configuration I had added:

```json
  "jest": {
    "testPathIgnorePatterns": [
      "/node_modules/",
      "/build/"
    ]
  }
```

Since I was unaware that these patterns were matching against the full
path, it took me a while to realize that `/build/` is part of the root
path for all Travis CI runs, and that this pattern was causing all of my
tests to be ignored.

We discussed this a little in #1047 and have opened #1057 to make
this scenario clearer, but I think that the better documentation would
also help prevent people from getting into this pickle in the first
place.

The solution for my problem was to add the `<rootDir>` string token to
my patterns, to make them include the root directory, preventing them
from matching, and ignoring, every file:

```json
  "jest": {
    "testPathIgnorePatterns": [
      "<rootDir>/node_modules/",
      "<rootDir>/build/"
    ]
  }
```

I tried to tailor the examples to each specific option, but as I'm not
that familiar with all of these configuration options, I may have missed
the mark on some of them.

I think that the defaults should perhaps also be changed to include the
`<rootDir>` string token as well, but I'll leave that for another
commit.