Add string replacements to preprocessor
Note: These changes are 100% the result of reading [this stackoverflow answer](http://stackoverflow.com/a/33024388/3990593) and all credit must really go to @TwitchBronBron

What these changes do is add istanbul ignore statements to the transpiled code so that the reported coverage better reflects reality on coverage reports. (Without this, the generated `__extends` statements mess with coverage reports)

Tested a few times and it seems to be working great.