Use `statSync` instead of `accessSync`.
Summary:
`accessSync` will return true for directories, which we do not want.

Right now, when in the integration test folder, running `jest snapshot` will not find any tests while `jest snapshot.js` will work. This is because there is a snapshot folder, `accessSync` will return true and Jest tries to run the directory as a test file.
Closes https://github.com/facebook/jest/pull/1101

Reviewed By: kentaromiura

Differential Revision: D3375163

fbshipit-source-id: f63809aab7397568b8dd7440a63cb44313f00a41