Warn for duplicate manual mock files (#1808)
* Warn for duplicate manual mock files

When a duplicate manual mock file is encountered, Jest silently takes
the last one that it happens to find when crawling the file system. This
can vary depending on the system, and there is no way for users to know
which mock was being used or that there are duplicates cluttering up
their project.

But have no fear ~ Now Jest will warn for the duplicate files, so that
you can delete duplicates and keep your projects free of dead code.

When running the tests for Jest itself, we had a fake set of 'modules'
that included a set of duplicate manual mocks for testing that very
behavior. The repeated set-up of the runtime with those fake modules
caused the warning to appear over and over, and it was too noisy.

The warning also conflicted with at least one other test that was
examining the warnings that should be thrown in another situation.

To minimize the noise, we moved the fake duplicate mock files to a
separate directory and only include them in the runtime for the test
that checks the behavior related to duplicate manual mocks..

This feels a bit hacky; suggestions are welcome for how to do this
better.

One idea is to expose an option to turn off all or some warnings thrown
by Jest, and we could use that in all tests which are not testing the
warning behavior itself.

Test plan:
- Unit tests pass
- Tested internally as well

* Remove extra 'package.json' file in dummy test files

This file was not being used, and was causing a warning to show when
running the tests;
```
    console.warn packages/jest-cli/node_modules/jest-runtime/node_modules/jest-haste-map/build/index.js:289
      jest-haste-map: @providesModule naming collision:
        Duplicate module name: my-module
        Paths: /Users/flarnie/Code/jest/packages/jest-runtime/src/__tests__/test_root/module_dir/my-module/package.json collides with /Users/flarnie/Code/jest/packages/jest-runtime/src/__tests__/test_root/subdir2/module_dir/my-module/package.json

      This warning is caused by a @providesModule declaration with the same name across two different files.
```