Avoid an infinite loop in `wrapAnsiString`
If a terminal's width is reported by node as `0`, this would get caught in an infinite loop, and crash v8 trying to allocate an ever-expanding array.

This works around this by immediately returning the supplied string if the width supplied is `0`, and adds a test for such a case.

fixes #1863