[Mocks] Make ModuleMocker create mock objects from the test environment
ModuleMocker was creating mock objects in the parent VM context but for consistency, we want to create them within the same environment as the one used to run the tests in a child VM context. This way if you create a mock object and compare its prototype with Object/Function/Array.prototype within a test, they will match.