Stop using Node `--inspect` option in the docs on debugging (#1998)
The technique described in the instructions does not pick up
breakpoints set with `debugger` statements. This issue is being
tracked by https://github.com/facebook/jest/issues/1652, but
it's unclear how long it will take to resolve as it's actually
a Node issue.

As a work-around, have updated docs to describe connecting an
external debugger to the Node process.