Always recreate mocks in FakeTimers.useFakeTimers()
If the flag resetMocks or jest.resetAllMocks is used in the test, all the timer
tests following that will fail with no way of recovering.

jest.resetAllMocks resets everything, including the one created by fakeTimers.
There is also no public methods that we can use to reinstantiate a new
FakeTimer object to mitigate that.

Given both mock and fake timers are part of the jest, they should be working
together out-of-the box without extra code gluing them. We change the
useFakeTimers method to always create new mocks so it would not be reverted.