Make jest-mock return an object
Right now it returns a class that you've got to instantiate which is not the API we want. The _ hack is ugly but I think we can live with it.

Test Plan:
Make sure tests are passing

Also make sure that we can write this :)
```js
var mock = require('jest-mock');
const fn = mock.getMockFn();
```