[Breaking change] Properly escape snapshots
There were many issues with snapshot test escaping:
- When escaping backtick, it didn't escape backslash
- Double quote escaping was overzealous and also escaped single quotes
- JSX text didn't use HTML escaping
- Regex escaping didn't escape all the control characters that should be

Because of the first one, you couldn't do eval(snapshot) and get the correct value, in order to make it kind of work, an unescape function was created that did ... something. Now that escaping is done properly we don't need it anymore.

This is a breaking change as all the double quotes now need to be escaped which triggers a ton of failures.