Removed `.mockComponent`
It looks like the default function exported from 'jest-react-native' is the `mockComponent` function. When I was running `jestReactNative.mockComponent`, I would get `TypeError: jestReactNative.mockComponent is not a function`. In my code, removing that and just running `jestReactNative('TouchableHighlight')` worked so I just removed it from the documentation to reflect that.