Initial commit: adds benchmarking functionality
This will add benchmarking functionality to Jest. The cli will detect benchmark files in __benches__ directories and pass them to BenchmarkRunner. It does not use the test running route, instead opting to use its own path to handle how benchmarks operate. Under the hood, the jest-runtime is re-used to ensure benchmarks are piped through the transforms correctly, opting to offer a new syntax in terms of: "benchmark", "scenario" for benchmark files. These do not use Jasmine, but has Jasmine like functionality. Benchmark.js is used to perform the operations per second on scanrios and a custom benchmark reporter outputs the results to the console.