Improve TS instructions (#3414)
The regex for TS transformer in the doc is too broad. For example, I set
the regex to `".(ts|tsx)"` as recommended in the docs and my test would
fail. The reason is that  one of my tested file imports the file `isArguments.js`
from `lodash-es` (ES 2015 build of lodash). Because `isArguments.js` matches the pattern
`.(ts|tsx)`, this file is not transpiled by babel (even after I've set
`transformIgnorePatterns` to `"<rootDir>/node_modules/(?!lodash-es)"`.
Changing the pattern to `"^.+\\.tsx?$"` fixes this problem.