Ability to fine-tune transform plugins via config
## Summary

If applied, this commit will enable to pass custom plugins to the babel transformer that jest uses internally. — The configuration is optional, therefore when the option is not present jest runtime will work as it was before.

## Why was this change made?

When I tried to use jest with a nested set of projects, I realized that some of the transformation that I defined on `.babelrc` (like, for example ` "plugins": [ "transform-flow-strip-types", "istanbul", "transform-es2015-modules-commonjs" ] ` were not being proppagated to the babel transformer that `jest` was using.

I set `transformIgnorePatterns`; however, that partially solved my problem since I wanted a different type of transformation for certain dependencies that I had.

So, to scratch my own itch, I extended to configuration to add that ability.

## Any side effects?

The patch is totally backwards-compatible: When the option is not present, the code works as it was before. And when the option is given, the code honors additional options.

I’ve run the whole test suite and everything is green.

## Example Configuration

package.json:

```json
  "jest": {
    "transformPlugins": [
      "transform-flow-strip-types",
      "istanbul",
      "transform-es2015-modules-commonjs"
    ]
  },
```