Fix incorrect toHaveProperty() documentation (#3676)
`toHaveProperty()` documentation incorrectly states value is be checked with strict equality (i.e. `===` like `toBe()`). It's actually checked with deep equality like `toEqual()`.

I've corrected the ExpectAPI documentation page to reflect this, and have made it absolutely clear which type of equality is used for this matcher.

Further suggestions: The page is quite inconsistent in how it talks about equality. It'd also be convenient to have an object property matcher that did use strict equality (e.g. `toHaveStrictProperty()` etc).