Remove dead reference
`jasmine.matchersUtil` is not included on this fork of jasmine, so `undefined` is being passed to the custom matcher. We make sure we remove that dead reference and we add a comment on top, so if people looks why they are not getting the matching utilities passed, they get a hint on how to proceed and don't get confused.