Don't format node assert errors when there's no 'assert' module
In format_node_assert_errors.js, we require the 'assert' module to
detect that an error thrown is an AssertionError, so that we can format
assertion errors nicely. This creates an implicit dependency in the
library on the assert module (a node builtin, but not available without
adding a shim in browsers, React Native, etc).

If we are running in an environment where assert is not available, we
don't need to try to format assertion errors, so we can just bail.
Detecting this case and performing the require in a try/catch removes
the implicit dependency on assert.

Fixes #4365