Change Promise detection code in jest-circus to support non-global Promise implementations (#4375)
* Change Promise detection code in jest-circus to support non-global Promise implementations

To provide better interop for Promise libraries such as bluebird and Q, this changes
the part of jest-circus that detects a returned Promise from a test or hook to only
check that the value is an object with a then method on it, rather than checking that
the value is instanceof Promise. This is considered the standard way of checking for
a Promises/A+-compliant Promise. As an added bonus, this check works cross-realm.

* Update utils.js