Allow custom resolver to be used with[out] moduleNameMapper (#4174)
* Allow the custom resolver to be used with and without the moduleNameMapper.

Also, improve error message when moduleNameMapper doesn't resolve to a module.

* Add 2nd test, to ensure custom resolver works without moduleNameMapper.

* Fix type and lint errors.

* Bug: rootDir wasn't the same value when --runInBand vs parallel

rootDir was modified after initialization to be process.cwd. In case of --runInBand,
The resolver was intialized before the rootDir value was altered. In case of parallel,
the Resolver was initialized in each spawned process using values of rootDir after
initialization.

Fix is to introduce a new config.cwd option that will be set to the parent processes
cwd. Adjust the console reporters to favor config.cwd over config.rootDir.

* React to changes upstream. Specify rootDir in defaultResolver.