Check if the maxWorkers has a value and raises an exception otherwise
In this commit we are adding a new validation in the check method of the
jest-cli package, so we raise an exception if the user tries to run Jest
with undefined maxWorkers, e.g.

jest --maxWorkers

jest --maxWorkers 2

Some users were confusing -w as an alias to --watch, but in fact it is
an alias to --maxWorkers. This change will also make this distinction
clearer.

Fixes issue #4577