jest-haste-map: add test to demo broken duplicates behavior (#4800)
`jest-haste-map` doesn't properly recover from duplicates when such state is stored in the cache (on the other hand, processing of duplicate by the watching mode is already correct and covered by tests). This changeset adds a test that demonstrate the brokeness. Next step will be to fix the behavior, and we can then switch the test to the correct values.