jest-haste-map: properly recover duplicates when files change (#4803)
When we load the haste map from the cache, we need to keep track of all the files that were removed, or even which module name changed, so that we can recover other modules with which there were potentially sharing names before.

For example, if we delete a module "Strawberry", maybe this allows another module that's the legitimate Strawberry to be available. Or, if we renamed a module's name from "Strawberry" to "Blackberry", maybe this allows another module to be deduplicated the same way.