Prevent a ENOENT crash by checking for file existence in jest-message-util (#5405)
* Prevent a ENOENT crash by checking for existence

Jest tries to load the stack-trace based on the top frame. The problem is, the filename may be provided by the source-map, which is unreliable since it can point to a non-existent file. In such a case, the code would cause a hard crash of "Test suite failed to run" with a stack-trace pointing at "exports.formatStackTrace", which I have corrected here.

* add an integration test for bad-source-map case

* lint CHANGELOG

* use a try-catch rather than fs.existsSync