Add some docs on using `jest.mock(...)`
I find the current docs on mocking lacking in the area of mocking an existing module. The hope is this example will show people a simple, real-world way to mock node modules (whether from npm or in their own projects). 

I looked through all the docs and couldn't find a clear example like this so I hope it is helpful to others!