Update TutorialReact.md (#5949)
* adds create-react-app snapshot dependency in docs

The current docs imply you can use Jest snapshots with create-react-app 'out-the-box', however the Snapshot Testing example doesn't work with a fresh install. You need to additionally install `react-test-renderer` as a dependency.

I thought it was more helpful to change the docs, but if it's better to raise the issue elsewhere let me know?!

* Update TutorialReact.md

The current docs imply you can use Jest snapshots with create-react-app 'out-the-box', however the Snapshot Testing example doesn't work with a fresh install. You need to additionally install `react-test-renderer` as a dependency.

I thought it was more helpful to change the docs, but if it's better to raise the issue elsewhere let me know?!

* ran yarn install and yarn lint:md

* amendned text to match w/o CRA code example