Add test to repro problem with mocking out 'getDerivedStateFromProps'
**what is the change?:**
Soon folks will be adopting the new 'getDerivedStateFromProps' in their
React components, and re-running existing tests.

If those components are mocked in existing tests, then React will throw
errors currently.

Why? Because React validates that any component defining this method has
state and returns the right thing from 'gDSFP', and when an auto-mock is
created it doesn't fit these invariants.

Talking through this with @mael and @mjesun we determined an easy
solution is to add a small conditional to either blacklist mocking
functions with this name, so that automocks just don't have this method
and the invariant is never checked in React.

An alternative solution would be to make the mocked 'gDSFP' method
return 'null' and define 'state' on the mocked component.

**why make this change?:**
Let's work together to make Jest and React work together incredibly
smoothly, like a sea of butter. ^_^

**test plan:**
This test will throw errors right now - tagging @mjesun to tweak Jest so
that the new test will pass correctly.

(Flarnie will insert screenshot of the errors thrown)