Update docs to detail resetModules: false behavior
The documentation at the moment describes the behavior of Jest when
`resetModules` is set to `true`, but leaves the default/false case
unclarified. I had a misconception that the module registry would not be
reset at all in the false case, where it's still reset between different
test files. This misconception was clarified by SimienB in issue #6007.

The reporter of issue #4413 also had this question.

This commit updates the documentation of `resetModules` to explain the
default case more explicitly.