Add help for configuration of collectCoverageFrom
Due to breaking changes in an upgrade of the dependency micromatch
from ^2.3.11 to ^3.1.10 many users saw their coverage reporting
failing because their glob matching was in an unsupported format.

Adding the help text here gives users a good starting point for
debugging when they run into issues rlated to this option. This will
help alleviate support issues (such as #6563) concerning this configuration
option.