Updates SnapshotTesting.md to provide more info. on snapshot scope
The doc doesn't mention where to look for the bugs in the code. I have added extra info. which will clear things to find the bugs and why does test pass when you doesn't add any props in the component used in different files.

As a beginner I thought, test will fail if I change Link component props in the any file that usage test component (in this case Link component).