Mock "exit" in "process"
Since functions exposed in `process` are the real ones, `exit` is being exposed as-is. The issue is that calling `process.exit` with a zero exit code might trick the workers, thinking the execution was stopped.

Pretty much like we do with `send`, we will override it with an empty method that does nothin.