Add package name to NotifyReporter notification (#5898)
## Summary

This change adds package name to notification in NotifyReporter when using `--notify` flag or `notify` configuration option. This clarifies which package is being tested which is especially useful when watching tests in mono-repos and with multiple IDEs opened.

First I'm trying to get hasteFS from the first context to get the module name (must be a better way). Then I'm falling back to `rootDir` and otherwise to original behavior without an additional title. 

Before:

![Before](https://user-images.githubusercontent.com/5013206/38090619-71194b18-3363-11e8-861e-4ca5566ea6a3.png)

After:

![After](https://user-images.githubusercontent.com/5013206/38090620-71388e10-3363-11e8-88fa-65e8ca605ef0.png)