Update Troubleshooting.md
With the existing launch configuration, VS Code automatically appends `--inspect-brk=<random_port>` resulting in a command like `node --inspect-brk ${workspaceRoot}/node_modules/.bin/jest --runInBand --inspect-brk=35370`. This does not work because the debugger is listening on port 9229, but Code is attempting to attach to it on port 35370.

Adding the port parameter prevents Code from automatically appending the `--inspect-brk` flag and instructs Code attach to the debugger on port 9229.