feat: symlinks in node_modules
- add `follow` method to `HasteFS` class
- add `links` property to `InternalHasteMap` type
- add `LinkData` map type
- add `LinkMetaData` array type

The user will need Watchman installed for this commit to take effect. Support for the Node crawler should be added in a future commit.

Only symlinks matching `node_modules/*` or `node_modules/@*/*` are stored in the `links` property. For each symlink found, a `[target, mtime]` tuple is created. The `target` value is undefined until `hasteFS.follow` uses `fs.realpathSync` to resolve the symlink. In this way, symlinks are lazily resolved and then cached.