Update TestingAsyncCode.md
The wording is confusing cause in one place it is "The problem is that the test will complete as soon as fetchData completes, before ever calling the callback." and another "Be sure to return the promise - if you omit this return statement, your test will complete before fetchData completes.". Logically both cannot be true and thus are confusing. The second statement is "worng", of course fetchData() does complete but the promise it returned is not necessarily resolved at this stage and the test will be finished ignoring the promise.