Fix custom async matcher stack trace
Fixes a bug causing custom async matcher stack traces to be lost. The
issue was caused by the JestAssertionError being created after the
promise for the async matcher resolved - by which point the stack
containing the correct stack trace (pointing to the line where matcher
was called) has been discarded. The issue was fixed by passing an error
that is initialized before the promise resolves.